Migrate to Modular Search
