+++ api/lib/search/arcgis-config.ts
